Display default empty template value if as_bootstrap can't render.
If the template code basses in an empty context variable, the
`bootstrap_input_type` templatetag throws a ValueError that's difficult
to debug.  This patch changes the template tag to just show an empty
string (or whatever is defined in TEMPLATE_STRING_IF_INVALID), so the
debugging can be done at the rendering level.